use python logging library and wait for user input before generating video

+++ src/agents/documentary_agent.py
@@ -4,6 +4,9 @@ from tools.video_stitcher import VideoStitcherTool
 from tools.wikipedia import WikipediaOnThisDayTool
 from chains.documentary_chain import NewsScriptChain
 from tools.video import NewsVideoGeneratorTool
+from logging import getLogger
+
+logger = getLogger(__name__)
 
 
 class OnThisDayDocumentaryAgent:
@@ -23,28 +26,38 @@ class OnThisDayDocumentaryAgent:
 
         script = self.chain.run(formatted)
 
-        print("****Generated script****")
-        print(script)
-        print("************************")
+        logger.info("\n****Generated script****")
+        logger.info(script)
+        logger.info("************************\n")
         return str(self.generate_long_video(script))
 
-    def generate_long_video(self, script: str) -> Path:
+    def generate_long_video(self, script: str) -> Path | None:
         chunks = script.split("\n\n")
         clip_paths: list[Path] = []
 
-        print(f"Generating long video ({len(chunks)} chunks)...")
+        if len(chunks) == 0:
+            logger.error("\n\nNo script chunks to process.\n\n")
+            return None
+
+        if len(chunks) < self.CHUNK_LIMIT:
+            logger.error("\n\nNot enough chunks to generate a long video.\n\n")
+            return None
+
+        logger.info(f"\nGenerating long video ({len(chunks)} chunks)...")
         for i, chunk in enumerate(chunks):
             if i >= self.CHUNK_LIMIT:
                 break
-            print(f"****Generating video for chunk {i + 1}****")
-            print(chunk)
+            logger.info(f"\n\n****Generating video for chunk {i + 1}****")
+            logger.info(chunk)
+            ans = input("Continue? (y/N): ")
+            if ans.lower() != "y":
+                return None
             video_id = self.video.generate_video(chunk)
             clip_path = Path(f"{self.outdir}/clip_{i}.mp4")
             self.video.download_video(video_id, clip_path)
             clip_paths.append(clip_path)
             self.video.delete_video(video_id)
-            print("*******************************")
-
+            logger.info("*******************************")
         final_path = Path(f"{self.outdir}/final_news_video.mp4")
         self.stitcher.stitch(clip_paths, final_path)
         return final_path
